backend part doe

+++ backend/controllers/goalController.js
@@ -1,5 +1,6 @@
 const asyncHandler = require('express-async-handler');
 const { request } = require('http');
+const User = require("../models/userModel")
 
 const Goal = require('../models/goalModel')
 
@@ -8,7 +9,8 @@ const Goal = require('../models/goalModel')
 // @route GET /api/goals
 // @access Priate
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json({ message: goals });
 })
@@ -24,7 +26,8 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json({ message: goal })
@@ -43,6 +46,20 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error("Goal not found")
     }
 
+    const user = await User.findById(req.user.id);
+
+    //Check for user exists
+    if (!user) {
+        res.status(401);
+        throw new Error("User not found");
+    }
+
+    //Make sure the logged user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error("User not authorized");
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true })
 
     res.status(200).json({ message: `Updated goal ${updatedGoal}` })
@@ -59,6 +76,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400);
         throw new Error("Goal not found");
+    }
+
+    const user = await User.findById(req.user.id);
+
+    //Check for user exists
+    if (!user) {
+        res.status(401);
+        throw new Error("User not found");
+    }
+
+    //Make sure the logged user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error("User not authorized");
     }
 
     const result = await Goal.findByIdAndDelete(req.params.id);
